Refactor integ tests and main code
